Cleaned up Point, Size and Rect API Separate clases for PointI/PointF Separate clases for SizeI/SizeF Separate clases for RectI/RectF

+++ include/Pt/Gfx/Api.h
@@ -55,7 +55,7 @@ namespace Pt {
 */
 namespace Gfx {
 
-typedef float Float;
+typedef double Float;
 typedef std::ptrdiff_t Int;
 
 class Canvas;
